Added rendering folder, moved Instance rendering scripts there
