--- Assets/Scripts/Cameras.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 530e45a1f5d532f418103e724b480ed2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
