+++ Assets/_Application/Rendering.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f1499d1a7f0444902b652db1efc29087
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
